+++ lib/pycarl/__init__.py
@@ -1,6 +1,7 @@
 from . import core
 from .core import *
 from . import infinity
+from . import _config
 
 __version__ = "unknown"
 try:
@@ -10,3 +11,37 @@ except ImportError:
     pass
 
 inf = infinity.Infinity()
+
+
+def carl_version():
+    """
+    Get Carl version.
+    :return: Version of Carl.
+    """
+    return _config.CARL_VERSION
+
+
+def has_cln():
+    """
+    Check if pycarl has support for CLN.
+    :return: True iff CLN is supported.
+    """
+    return _config.CARL_WITH_CLN
+
+
+def has_parser():
+    """
+    Check if pycarl has parsing support.
+    :return: True iff parsing is supported.
+    """
+    return _config.CARL_WITH_PARSER
+
+
+def print_info():
+    """
+    Print information about pycarl.
+    """
+    print("Pycarl version {}".format(__version__))
+    print("Using carl in version {}".format(carl_version()))
+    print("Support for CLN: {}".format(has_cln()))
+    print("Support for parsing: {}".format(has_parser()))