+++ tests/configurations.py
@@ -1,12 +1,10 @@
 import pytest
 
-import pycarl._config as config
+import pycarl
 
 # Skip not supported functionality
-has_cln = config.CARL_WITH_CLN
-has_parser = config.CARL_WITH_PARSER
-cln = pytest.mark.skipif(not has_cln, reason="No support for CLN")
-parser = pytest.mark.skipif(not has_parser, reason="No support for carlparser")
+cln = pytest.mark.skipif(not pycarl.has_cln(), reason="No support for CLN")
+parser = pytest.mark.skipif(not pycarl.has_parser(), reason="No support for carlparser")
 
 # Parametrize available number types
 import pycarl.gmp
@@ -14,7 +12,7 @@ import pycarl.gmp
 parameters = [pycarl.gmp]
 names = ["gmp"]
 
-if has_cln:
+if pycarl.has_cln():
     import pycarl.cln
 
     parameters.append(pycarl.cln)