+++ tests/convert/conftest.py
@@ -1,9 +1,9 @@
-from configurations import has_cln
+import pycarl
 import pycarl.convert
 import pycarl.formula
 import pycarl.gmp
 import pycarl.gmp.formula
 
-if has_cln:
+if pycarl.has_cln():
     import pycarl.cln
     import pycarl.cln.formula