+++ tests/convert/test_convert_python_types.py
@@ -1,12 +1,10 @@
 import pytest
 import pycarl
 
-from configurations import has_cln
-
 parameters = [(pycarl.gmp, pycarl.convert.gmp_converter)]
 names = ["gmp_converter"]
 
-if has_cln:
+if pycarl.has_cln():
     parameters.append((pycarl.cln, pycarl.convert.cln_converter))
     names.append("cln_converter")