+++ tests/formula/conftest.py
@@ -1,5 +1,5 @@
-from configurations import has_cln
+import pycarl
 import pycarl.gmp.formula
 
-if has_cln:
+if pycarl.has_cln():
     import pycarl.cln.formula